CHROMIUM: dm_verity: speed up boot when not using verity
If we do not want to wait for verity to become available and we can't
fain the device there is no point in waiting 500 msecs.

Also, if we fail to find the verity device we should not be trying to
perform async_schedule_full(): for us it is too later and upper layers
take care of waiting/synchronizing on their own.

BUG=None
TEST=Build and boot developer image, updated kernel.

Change-Id: I5d04239d263e0b3c4acb3c7a602dee7406f90463
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241122
Reviewed-by: Olof Johansson <olofj@chromium.org>