boot_constraint: Add support for Hisilicon platforms
This adds boot constraint support for Hisilicon platforms. Currently
only one use case is supported: earlycon. One of the UART is enabled by
the bootloader and is used for early console in the kernel. The boot
constraint core handles it properly and removes constraints once the
serial device is probed by its driver.

This is tested on hi6220-hikey 96board.

Cc: Wei Xu <xuwei5@hisilicon.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>