clk: Remove cached cores in parent map during unregister
As clocks are registered their parents are resolved and the parent_map
is updated to cache the clk_core objects of each existing parent.
But in the event of a clock being unregistered this cache will carry
dangling pointers if not invalidated, so do this for all children of the
clock being unregistered.

Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>