clk: qcom: gcc-sdm845: Add clocks to enable access to some qfprom regions
On SDM845, access to certain qfprom regions requires additional clocks to
be turned on. This patch models the RCG and the branch clock needed to enable
access to those regions.

These clocks are marked as *read only* on some qualcomm firmware and can cause
an abort if HLOS tries to write into these clock control registers. So these
need to be marked as protected-clocks on those platforms (like MTP)

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>