INTERNAL: arm64: Add support for NMI CPU backtrace
Hardlockup debugging with the current backtrace is almost
useless as the same backtrace for any hardlockups are seen
due to missing NMI CPU backtracing support on arm64. But
now that SGIs can be handled as pseudo NMIs, we can use
it to have a arch_trigger_cpumask_backtrace() similar to
arm32 for IPI based dumping of CPU stacks which will give
the stack for the locked CPU as well and ease debugging.

We will need CONFIG_HARDLOCKUP_DETECTOR_BUDDY_CPU=y and
CONFIG_ARM64_PSEUDO_NMI=y and irqchip.gicv3_pseudo_nmi=1.

Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>